_archive/van-controller_legacy/CORE_project/requirements.txt
Flask
requests
paho-mqtt
influxdb-client
